refactor: component group into pages.
