--- src/components/Footer.astro
@@ -1,122 +1,0 @@
----
-import Link from "@/components/Link.astro"
-import Svg from "@/components/Svg.astro";
-
-const { i18n } = Astro.locals;
----
-
-<footer class="bg-black/40">
-  <div class="mx-auto w-full max-w-screen-xl p-4 px-6 py-6 lg:py-8">
-    <div class="md:flex md:justify-between">
-      <div class="mb-6 flex flex-col md:mb-0">
-        <a
-          href="https://premiosesland.com/"
-          class="-mt-14 flex items-center justify-center"
-        >
-          <Svg src="logo" class="h-auto w-24" />
-        </a>
-        <div class="mt-6 flex justify-center gap-x-4">
-          <a href="#" class="text-white hover:text-primary duration-300">
-            <Svg src="twitter" class="size-6" hidden />
-            <span class="sr-only">Twitter</span>
-          </a>
-          <a href="#" class="text-white hover:text-primary duration-300">
-            <Svg src="instagram" class="size-6" width="256" height="256" />
-            <span class="sr-only">Instagram</span>
-          </a>
-          <a href="#" class="text-white hover:text-primary duration-300">
-            <Svg src="reddit" class="size-6" width="24" height="24" />
-            <span class="sr-only">Reddit</span>
-          </a>
-        </div>
-      </div>
-      <div class="grid grid-cols-2 gap-8 sm:grid-cols-3 sm:gap-6">
-        <div>
-          <h2 class="mb-6 text-sm font-semibold uppercase text-white">
-            {i18n.FOOTER.LEGAL.TITLE}
-          </h2>
-          <ul class="flex flex-col gap-2 font-medium text-gray-400">
-            <li>
-              <Link
-                href="/aviso-legal"
-                class="hover:underline hover:text-primary duration-300"
-                >{i18n.FOOTER.LEGAL.NOTICE}</Link
-              >
-            </li>
-            <li>
-              <Link
-                href="/privacidad"
-                class="hover:underline hover:text-primary duration-300"
-                >{i18n.FOOTER.LEGAL.PRIVACY}</Link
-              >
-            </li>
-            <li>
-              <Link
-                href="/cookies"
-                class="hover:underline hover:text-primary duration-300"
-              >
-                {i18n.FOOTER.LEGAL.COOKIES}
-              </Link>
-            </li>
-          </ul>
-        </div>
-        <div>
-          <h2 class="mb-6 text-sm font-semibold uppercase text-white">
-            {i18n.FOOTER.WEB_MAP.TITLE}
-          </h2>
-          <ul class="flex flex-col gap-2 font-medium text-gray-400">
-            <li>
-              <Link
-                href="/"
-                class="hover:underline hover:text-primary duration-300"
-                >{i18n.FOOTER.WEB_MAP.HOME}</Link
-              >
-            </li>
-            <li>
-              <Link
-                href="/info"
-                class="hover:underline hover:text-primary duration-300"
-                >{i18n.FOOTER.WEB_MAP.INFO}</Link
-              >
-            </li>
-            <li>
-              <Link
-                href="/archivo"
-                class="hover:underline hover:text-primary duration-300"
-                >{i18n.FOOTER.WEB_MAP.ARCHIVE}</Link
-              >
-            </li>
-          </ul>
-        </div>
-        <div>
-          <h2 class="mb-0 text-sm font-semibold uppercase text-white">
-            {i18n.FOOTER.PRODUCED_BY}
-          </h2>
-          <a
-            class="hover:opacity-100 opacity-70 transition-opacity duration-300 block"
-            href="https://www.ubiq.ad/"
-            target="_blank"
-            rel="noreferrer"
-          >
-            <Svg src="ubiq-logo" label="Logo Ubiq" />
-          </a>
-        </div>
-      </div>
-    </div>
-    <hr class="my-6 border-white/10 sm:mx-auto lg:my-8" />
-    <div class="sm:flex sm:items-center sm:justify-between">
-      <span class="text-sm text-gray-400 sm:text-center"
-        >{i18n.FOOTER.DISCLAIMER}
-      </span>
-      <div class="mt-4 flex sm:mt-0 sm:justify-center">
-        <a
-          href="https://github.com/midudev/esland-web"
-          class="text-gray-500 hover:text-white ms-5 duration-300"
-        >
-          <Svg src="github" class="h-4 w-4" hidden />
-          <span class="sr-only">GitHub account</span>
-        </a>
-      </div>
-    </div>
-  </div>
-</footer>
