+++ src/components/Hero.astro
