+++ src/components/archivo/Galeria.astro
