+++ src/components/archivo/Numeros.astro
@@ -1,6 +1,6 @@
 ---
-import { CountUp } from "@/components/CountUp"
-import NumerosItem from "@/components/NumerosItem.astro"
+import { CountUp } from "@/components/archivo/CountUp"
+import NumerosItem from "@/components/archivo/NumerosItem.astro"
 
 const { i18n } = Astro.props
 ---
