+++ src/components/archivo/Palmares.astro
@@ -1,5 +1,5 @@
 ---
-import StreamerCard from "@/components/StreamerCard.astro"
+import StreamerCard from "@/components/archivo/StreamerCard.astro"
 import palmares from "@/content/editions/info.json"
 
 interface Props {
