+++ src/components/info/InfoKitPrensa.astro
@@ -1,5 +1,5 @@
 ---
-import InfoKitPrensaItem from "@/components/InfoKitPrensaItem.astro"
+import InfoKitPrensaItem from "@/components/info/InfoKitPrensaItem.astro"
 const { i18n } = Astro.locals
 ---
