+++ src/components/pages/Inicio.astro
@@ -1,9 +1,9 @@
 ---
 import Layout from "@/layouts/Layout.astro"
-import Video from "@/components/Video.astro"
+import Video from "@/components/inicio/Video.astro"
 import Intro from "@/components/Intro.astro"
-import Bento from "@/components/Bento.astro"
-import PreFooter from "@/components/PreFooter.astro"
+import Bento from "@/components/inicio/Bento.astro"
+import PreFooter from "@/components/inicio/PreFooter.astro"
 
 const { i18n } = Astro.locals;
 ---
