+++ src/components/vota/VoteFinal.tsx
